requirements.txt
streamlit
google-generativeai
markdown
xhtml2pdf